requirementsLocal.txt
amqp==5.0.6
argon2-cffi==21.1.0
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==21.1
et-xmlfile==1.1.0
idna==2.10
kombu==5.1.0
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.3
Pillow==8.3.2
prompt-toolkit==3.0.21
psycopg2==2.9
psycopg2-binary==2.9
pyarrow==5.0.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==3.0.1